requirements.txt
alembic==1.15.2
bidict==0.23.1
blinker==1.9.0
cffi==1.17.1
click==8.1.8
cryptography==44.0.3
dnspython==2.7.0
eventlet==0.40.2
Flask==3.1.0
flask-cors==5.0.1
Flask-Migrate==4.1.0
Flask-SocketIO==5.5.1
Flask-SQLAlchemy==3.1.1
greenlet==3.2.3
h11==0.16.0
iniconfig==2.1.0
itsdangerous==2.2.0
Jinja2==3.1.6
Mako==1.3.10
MarkupSafe==3.0.2
packaging==25.0
pluggy==1.5.0
pycparser==2.22
pytest==8.3.5
python-dotenv==1.1.0
python-engineio==4.12.0
python-socketio==5.13.0
simple-websocket==1.1.0
SQLAlchemy==2.0.40
typing_extensions==4.13.2
Werkzeug==3.1.3
wsproto==1.2.0
